Remove unused deterministicId in AbstractSerializableSignatureParameters.
Its value is never changed from the default value of null.
Another deterministicId is defined in AbstractSignatureParameters which is also used and is accessible by other classes.

Note: there are spurious test failures in
  TwoPAdESSigniatureMustHaveDifferentIdTest.test:77 Values should be different. Actual: id-7c30b25092897a417fea88c5e5d39629a6b1eadf11b4ae2fb6a9085cc9c1744f
but this also happens on current version of the development branch when building the dss-pades package.