Fix for SunPKCS11 in Java 10 changes.
SunPKCS11 does not longer supports constructor with InputStream. Instead it uses method configure( String ).
This fix invokes the method through reflection. In future Java version it won't be possible so the library must be build in Java 11.
Now it is possible to build it on Java 8 and run on 9 and 10.