DSS-2609: Allow signature with external CMS provider
Motivation:
We want to perform PAdES signature with an external CMS provider. The current signing operation re computed the signedAttribute which cause the signature being invalid.
The ability to use the PAdESService is already supported in the DSS but the method are mainly package protect.
The goal here is just to turn the right method to public

Modification:
* PAdESService#computeDocumentDigest is now public
* PAdESService#signDocument is now based on PAdESService#signDocumentWithCms
* PAdESService#generateCMSSignedData is now public
* Simplify test with the methods