Simplified "toString" to avoid synchronization
StringWriter internally uses a StringBuffer, which uses synchronization over an internal object.
In this case the synchronization is not necessary, so should a simpler string composition is more efficient and effective. 
Furthermore, composing the result with a StringWriter prevents the compiler to use the String composition based on the invokedynamic implementation, that generate the CallSite only if necessary, and mask the fact that the final length can be calculated at creation, thus avoiding resizing and relocations.