DSS-2930 Decrease memory usage by returning byte array of the raw stream
Replace createInputStream() with createRawInputStream() when returning
a byte array for a COS stream in PdfBoxArray to avoid excessive memory
usage.
Using createInputStream creates internally an object presentation of
the content of a stream, e.g. an image will be rendered into memory,
and then returns the byte array of  the rendered representation.
This is not needed for the usage of getStreamBytes() in the
DSSDictionaryExtractionUtils for extracting certificates. There are
currently no other usages which needs a rendered respresentation.

The byte length of the rendered representation can be of a magnitude
bigger than the byte length of the raw stream, hence using
createRawInputStream(), which returns the stream as-is, greatly
decreases memory usage.